fix(AnswerFormula): calculate 메서드 반환 타입 int로 변경
- index로 사용되는 용도이므로 변경

+++ src/main/java/wordle/domain/AnswerFormula.java
@@ -3,5 +3,5 @@ package wordle.domain;
 @FunctionalInterface
 public interface AnswerFormula {
 
-    long calculate(int wordCount);
+    int calculate(int wordCount);
 }
